requirements.txt
flask==2.0.1
werkzeug==2.0.2
python-dotenv==0.19.0
openai==0.28.1
gunicorn==20.1.0